pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reefwatch"
version = "1.0.0"
description = "Continuous local security monitoring for OpenClaw hosts"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "OpenClaw Contributors" },
]
keywords = ["security", "hids", "intrusion-detection", "openclaw"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
]
dependencies = [
    "yara-python>=4.5.0",
    "pySigma>=0.11.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.scripts]
reefwatch = "reefwatch.daemon:main"

[project.urls]
Homepage = "https://github.com/openclaw/reefwatch"
Repository = "https://github.com/openclaw/reefwatch"
Issues = "https://github.com/openclaw/reefwatch/issues"

[tool.setuptools.packages.find]
include = ["reefwatch*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
